refactor(core): rename internal io and ui packages to utils/io and utils/ui
This change organizes internal packages by moving `internal/io` and
`internal/ui` under a new `utils` directory to better reflect their
utility nature.  All references have been updated accordingly.

No breaking changes for end users; this is purely an internal
refactoring.  ```
